Internal Data Audit Form
This form provides authors with a simple form for assisting with an internal data audit. Having a colleague check the raw data and/or analyses facilitates reproducibility and may prevent or detect data falsification or fabrication before a paper is submitted for publication. Publishing documentation of the quality control process provides assurance for readers and allows the work of reproducibility reviewing to be recognised.